Changed Model Layer And Test KAKAO_OAUTH

+++ src/main/java/com/baylab/project/domain/service/UserStorage.java
@@ -1,9 +1,10 @@
 package com.baylab.project.domain.service;
 
+import com.baylab.project.domain.model.Problem;
 import com.baylab.project.domain.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface UserStorage extends JpaRepository<User,Integer> {
-
+public interface UserStorage extends JpaRepository<User,Long> {
+    
 }
